[ticket/13782] Rename null cache driver to dummy
The null driver still exists as a backwards compatible class that extends
the dummy driver.

PHPBB3-13782
